+++ NEWS.md
@@ -1,6 +1,13 @@
 # MSnbase 2.7
 
-## Changes in version 2.7.1
+## Changes in version 2.7.4
+- Nothing yet 
+
+## Changes in version 2.7.3
+- Fix bug in robust summary (see PR [#349](https://github.com/lgatto/MSnbase/issues/349)) <2018-07-28 Sat>
+- Fix failing unit test <2018-07-28 Sat>
+
+## Changes in version 2.7.2
 - Handle files without any spectra - see
   [#342](https://github.com/lgatto/MSnbase/issues/342)
   <2018-05-15 Tue>
@@ -8,6 +15,9 @@
   <2018-05-30 Wed>
 - Update plot,Spectrum methods to match the tolerance and relative
   arguments (see [#350](https://github.com/lgatto/MSnbase/issues/350)) <2018-06-29 Fri>
+
+## Changes in version 2.7.1
+- Version bump to force new vignette build
 
 ## Changes in version 2.7.0
 - New devel version for Bioc 3.8